fix: remove default argument `now` from `_seconds_until_refresh`  (#356)
Having now as a default argument means it is set once and then never
adjusted. So in subsequent calls to _seconds_until_refresh the duration
will just continue to grow larger and larger as it evaluates agains the same now

The result of this means that duration // 2 block is always being hit and
will continue to grow. It could grow to be greater than 3600 and lead to errors.