chore: shorten duration of connector unit tests
test_connector.py and test_async_connector.py were taking around a
minute to run. The reason for this is that the AlloyDB Admin API client
sets a default deadline of 60s to retry failed API requests. This
increases the time for certain unit tests, which send actual API
requests. This PR overrides the timeout to 1s for those unit tests, to
speed up their execution.