fix: support verify-full mode (#446)
This PR allows the Python connector to support the most secure form of
mTLS, which is `verify-full` mode:
https://www.postgresql.org/docs/current/libpq-ssl.html#LIBQ-SSL-CERTIFICATES.
In `verify-full` mode, the hostname provided by the client is matched
with the AuthProxy server's certificate's SAN field. If it does not
match, the connection is rejected.

Previously, for PSC instances, the AuthProxy server certificates
contained DNS names with trailing dots in the SAN field. But Python's
OpenSSL library does not allow DNS names to contain trailing dots when
doing the hostname verification. So we could not support `verify-full`
mode.

As of now, all PSC instances that use the AlloyDB Python connector do
not have DNS names with trailing dots in the SAN field anymore. So we
can support `verify-full` mode now.